requirements.txt
snowflake-sqlalchemy
snowflake-connector-python[secure-local-storage,pandas]
pandas
jupyterlab
python-dotenv